Remove redundancies flagged by Kotlin 2.4.
Drop redundant toLong()/toInt()/toDouble() conversions, elvis operators
whose right-hand side is null or can never be reached, an always-true null
check, a redundant else branch in an exhaustive when, an unnecessary !!,
and use Any() rather than java.lang.Object() for a lock.

+++ jvb/src/main/kotlin/org/jitsi/videobridge/cc/BandwidthProbing.kt
@@ -62,7 +62,7 @@ class BandwidthProbing(
     var diagnosticsContext: DiagnosticContext? = null
 
     override fun bandwidthEstimationChanged(newValue: Bandwidth) {
-        latestBwe = newValue.bps.toLong()
+        latestBwe = newValue.bps
     }
 
     override fun run() {
